WepApi Part2

+++ Entities/Concrete/Bank.cs
@@ -1,14 +1,15 @@
 ﻿using Core.Entities;
+using System;
+using System.ComponentModel.DataAnnotations;
 
 namespace Entities.Concrete
 {
     public class Bank : IEntity
     {
-        
         public int BankId { get; set; }
-        public int GoldCoin { get; set; }
-        public int SilverCoin { get; set; }
-        public int CopperCoin {get; set;}
+        public Int16 GoldCoin { get; set; }
+        public Int16 SilverCoin { get; set; }
+        public Int16 CopperCoin {get; set;}
 
 
     }
